+++ packages/ndk/lib/domain_layer/entities/account.dart
@@ -1,0 +1,12 @@
+import '../repositories/event_signer.dart';
+
+enum AccountType { privateKey, publicKey, externalSigner }
+
+/// Account enitity
+class Account {
+  final AccountType type;
+  final String pubkey;
+  final EventSigner signer;
+
+  Account({required this.type, required this.pubkey, required this.signer});
+}
