+++ packages/ndk/lib/presentation_layer/ndk_config.dart
@@ -3,7 +3,6 @@ import '../config/logger_defaults.dart';
 import '../config/request_defaults.dart';
 import '../domain_layer/entities/event_filter.dart';
 import '../domain_layer/repositories/cache_manager.dart';
-import '../domain_layer/repositories/event_signer.dart';
 import '../domain_layer/repositories/event_verifier.dart';
 import 'package:logger/logger.dart' as lib_logger;
